+++ Sources/StreamChat/WebSocketClient/WebSocketClient_Tests.swift
@@ -590,6 +590,27 @@ class WebSocketClient_Tests: StressTestCase {
         // Check the background task is terminated
         AssertAsync.willBeEqual(backgroundTaskScheduler.endBackgroundTask_called, task)
     }
+    
+    func test_backgroundTaskIsCancelled_whenExpirationHandlerIsCalled() {
+        // Simulate connection and start a background task
+        test_connectionFlow()
+        let task = UIBackgroundTaskIdentifier(rawValue: .random(in: 1...100))
+        backgroundTaskScheduler.beginBackgroundTask = task
+        NotificationCenter.default.post(name: UIApplication.didEnterBackgroundNotification, object: nil)
+        
+        // Wait for `beginBackgroundTask` being called since it can be done asynchronously
+        AssertAsync.willBeTrue(backgroundTaskScheduler.beginBackgroundTask_called)
+        assert(backgroundTaskScheduler.endBackgroundTask_called == nil)
+        
+        // We don't simulate explicit cancelation here
+        // since we expect expiration handler to call disconnect
+        
+        // Call expiration handler
+        backgroundTaskScheduler.beginBackgroundTask_expirationHandler!()
+        
+        // Check the background task is terminated
+        AssertAsync.willBeEqual(backgroundTaskScheduler.endBackgroundTask_called, task)
+    }
 }
 
 final class HealthCheckMiddleware_Tests: XCTestCase {